lora-tx/requirements.txt
python-dotenv
pyserial
lgpio
